organize driver project structure
