--- HandOfLesserDriver/HandOfLesserDriver/CMakeLists.txt
@@ -1,45 +1,0 @@
-set(TARGET_NAME handoflesser)
-
-# This is the name of the driver according to SteamVR
-set(DRIVER_NAME "driver_${TARGET_NAME}")
-
-project(${TARGET_NAME})
-
-add_library(${DRIVER_NAME} SHARED
-    src/hmd_driver_factory.cpp
-    src/device_provider.cpp
-    src/emulated_controller_driver.cpp
-    src/hand_simulation.cpp
-    src/hooked_controller.cpp
-    src/hand_of_lesser.cpp
-    src/controller_common.cpp
-)
-
-message( "\${CMAKE_RUNTIME_OUTPUT_DIRECTORY}: ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}" )
-message( "\${TARGET_NAME}: ${TARGET_NAME}" )
-message( "\${ARCH_TARGET}: ${ARCH_TARGET}" )
-message( "\${OPENVR_LIBRARIES}: ${OPENVR_LIBRARIES}" )
-
-# This is so we can build directly to "<binary_dir>/<target_name>/<platform>/<arch>/<driver_name>.<dll/so>"
-set_target_properties(${DRIVER_NAME} PROPERTIES
-    RUNTIME_OUTPUT_DIRECTORY $<1:${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${TARGET_NAME}/bin/${ARCH_TARGET}>
-)
-
-target_link_libraries(${DRIVER_NAME} PRIVATE
-    ${OPENVR_LIBRARIES}
-    util_driverlog
-    util_vrmath
-    eigen
-    minhook
-    HandOfLesserCommon
-)
-
-target_include_directories(${DRIVER_NAME} PRIVATE ${OPENVR_INCLUDE_DIR})
-
-# Copy driver assets to output folder
-add_custom_command(TARGET ${DRIVER_NAME}
-    PRE_BUILD
-    COMMAND ${CMAKE_COMMAND} -E copy_directory
-        ${CMAKE_CURRENT_SOURCE_DIR}/${TARGET_NAME}
-        ${CMAKE_RUNTIME_OUTPUT_DIRECTORY}/${TARGET_NAME}
-)
