+++ HandOfLesserDriver/handoflesser/resources/localization/localization.json
