+++ HandOfLesserDriver/src/core/hand_of_lesser.cpp
